Add sqlite logging
Open a database from file given by command line argument -d/--db and add links
to this database.

There may be a better way to avoid having to open a database in memory if a
database is not required, for which there would presumably be some overhead.

An empty reference would have to be passed as something with a movable trait,
since it must be referenced inside a FnMut closure.